ml-service/requirements.txt
fastapi 
uvicorn 
joblib 
pandas 
scikit-learn==1.6.1
python-dotenv